fix: remove 5000ms timeout
We have issues both locally and in deployed environments where this timeout is too low. This makes the app unable to be deployed. Instead of setting some arbitrary higher timeout, we should remove it until we are able to set a reasonable timeout.

When we add minification to the bundling the time will go up even more.