feat: update to support typescript
Instead of determining the type of asset based on file extension, just say that it is the tinh it is sent is as.

Background: I'd like to use browserify and tsify to compile/transpile my ts/tsx files to js as they are bundled, but sending in a file with extension `.tsx` makes stuff break currently.

Also, changing order of plugins/transforms to run plugins before transforms, as it's mentioned here in the FAQ that it is required for at least tsify: https://www.npmjs.com/package/tsify